Changed file open logic to attempt lookup of relative paths by searching from the script that contains the include statement. Also fixed the lex stream logic; it wasn't properly adjusting the marker pointer. Upgraded to re2c trunk version.

+++ trunk/include/lexstream.h
@@ -58,7 +58,7 @@ public:
   //
   // Fill buffer and adjust pointers, called by re2c
   //
-  size_t FillBuffer(size_t minSize = 0);
+  int FillBuffer(int minSize = 0);
 
 private:
